==> dev/08_Visualization_Base.R <==
#-----------------------------------------------------------------------------#
# Visualization with Base R
#-----------------------------------------------------------------------------#

# Try to re-purpose the chapter from metrics book
# Re-org, re-word, but use all the same content 


# Options and controls ----------------------------------------------------


# Variation within variables ----------------------------------------------

# Discrete

# Continuous 


# Co-variation between variables ------------------------------------------

# Discrete

# Continuous
